requirements.txt
selenium
undetected-chromedriver
easyocr
python-dotenv
